# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaja (Kantin Jawara) API
  description: |
    # Kaja API Documentation

    API for **Kaja (Kantin Jawara)** canteen management system - a comprehensive solution for restaurant and food court operations.

    ## 🚀 Features

    - **Authentication System** - Secure user registration, login, and profile management
    - **Multi-Role Support** - Admin, Seller (Penjual), and Buyer (Pembeli) roles
    - **Menu Management** - Complete CRUD operations for food items and categories
    - **Smart Cart System** - Automatic price calculations, merchant-specific carts, and real-time updates
    - **Transaction Processing** - Order creation, status tracking, and management
    - **Payment Integration** - Multiple payment methods (Cash, Bank Transfer, E-Wallet)
    - **Real-time Chat** - Transaction-based messaging between buyers and sellers
    - **File Upload Support** - Images and documents for chat and payment proof

    ## 🛒 Cart Management

    The cart system provides intelligent shopping cart functionality with automatic calculations:

    - **Merchant-specific carts** - Separate carts for each merchant/seller
    - **Automatic price calculation** - Total prices calculated automatically based on quantity and unit price
    - **Real-time updates** - Cart totals recalculated instantly when items are added, updated, or removed
    - **Smart item management** - Duplicate items are merged with updated quantities
...

    ## 📋 Changelog

...

...

...

    ```
...
    ```

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /cart/{merchantId}:
    get:
      tags:
        - 🛒 Customer APIs
      summary: Get cart for specific merchant
      description: Get cart for a specific merchant
      operationId: getCartByMerchant
      parameters:
        - name: merchantId
          in: path
          required: true
          description: Merchant ID
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cart retrieved successfully"
                  data:
                    $ref: "#/components/schemas/Cart"
                  total_items:
                    type: integer
                    example: 3
                  total_amount:
                    type: number
                    format: decimal
                    example: 75000
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    delete:
      tags:
        - 🛒 Customer APIs
      summary: Clear cart for specific merchant
      description: Clear cart for a specific merchant
      operationId: clearCartByMerchant
      parameters:
        - name: merchantId
          in: path
          required: true
          description: Merchant ID
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cart cleared successfully"
                  cleared_carts:
                    type: integer
                    example: 1
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoint.

        **Usage:** `Authorization: Bearer your_token_here`
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                message: "Unauthenticated"
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          description: Category ID
          example: 1
        name:
          type: string
          description: Category name
          example: "Makanan Berat"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
...
    Menu:
      type: object
      properties:
        id:
          type: integer
          description: Menu ID
          example: 1
        name:
          type: string
          description: Menu item name
          example: "Nasi Goreng Spesial"
        description:
          type: string
          nullable: true
          description: Menu item description
          example: "Nasi goreng dengan telur dan ayam"
        price:
          type: number
          format: float
          description: Price in IDR
          example: 25000
        stock:
          type: integer
          description: Available stock
          example: 10
        image_url:
          type: string
          nullable: true
          description: Menu item image URL
          example: "https://example.com/nasigoreng.jpg"
        category:
          $ref: "#/components/schemas/Category"
        merchant:
          type: object
          properties:
            id:
              type: integer
              example: 2
            name:
              type: string
              example: "Warung Pak Joni"
        is_available:
          type: boolean
          description: Whether item is available (stock > 0)
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Something went wrong"
...
    Cart:
      type: object
      description: |
        Shopping cart containing items from a specific merchant.

        **Automatic Recalculation Events:**
        - When cart items are added, updated, or removed
        - When item quantities or prices change
        - When tax or discount amounts are modified

        The cart automatically maintains consistency between total_amount, tax_amount, 
        discount_amount, and final_amount through model events.
      properties:
        id:
          type: integer
          example: 1
          description: Unique cart identifier
        user_id:
          type: integer
          example: 2
          description: ID of the user who owns this cart
        session_id:
          type: string
          nullable: true
          example: "sess_abc123"
          description: Session identifier for guest users
        total_amount:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 75000.00
          description: Total amount of all items in the cart
        tax_amount:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 0.00
          description: Tax amount applied to the cart
        discount_amount:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 0.00
          description: Discount amount applied to the cart
        final_amount:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 75000.00
          description: Final amount after tax and discount calculations
        status:
          type: string
          enum: [active, completed, abandoned]
          example: "active"
          description: Current status of the cart
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
          description: List of items in the cart
        merchant:
          type: object
          properties:
            id:
              type: integer
              example: 2
              description: Merchant ID
            name:
              type: string
              example: "Warung Pak Joni"
              description: Merchant name
            email:
              type: string
              example: "penjual@example.com"
              description: Merchant email
          description: Information about the merchant for this cart
        total_items:
          type: integer
          description: Total quantity of items in cart
          example: 3
        created_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
          description: When the cart was created
        updated_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
          description: When the cart was last updated
      required:
        - id
        - user_id
        - total_amount
        - final_amount
        - status
    CartItem:
      type: object
      description: |
        Individual item within a shopping cart with automatic price calculation.

        **Automatic Calculations:**
        - total_price is automatically calculated as quantity × unit_price
        - Cart totals are recalculated when cart items change
        - Price calculations happen on create, update, and delete events
      properties:
        id:
          type: integer
          example: 1
        cart_id:
          type: integer
          example: 1
          description: ID of the cart this item belongs to
        menu_id:
          type: integer
          example: 5
          description: ID of the menu item
        quantity:
          type: integer
          example: 2
          minimum: 1
          maximum: 100
          description: Quantity of the menu item
        unit_price:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 25000.00
          description: Price per unit of the menu item
        total_price:
          type: number
          format: decimal
          multipleOf: 0.01
          example: 50000.00
          description: Total price for this cart item (quantity × unit_price)
        notes:
          type: string
          nullable: true
          maxLength: 255
          example: "Extra spicy, no onions"
          description: Special instructions or notes for this item
        menu:
          $ref: "#/components/schemas/Menu"
          description: The menu item details
        created_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
          description: When the cart item was created
        updated_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
          description: When the cart item was last updated
      required:
        - id
        - cart_id
        - menu_id
        - quantity
        - unit_price
        - total_price

    # Merchant Payment Method schemas
...